refactor: move manual sync logger to logger module
