+++ src/logging/logger.ts
@@ -1,4 +1,6 @@
-import { Request, Response } from 'express';
+import { writeFileSync } from 'fs';
+import { resolve } from 'path';
+import { Request, Response, NextFunction } from 'express';
 import { finished } from 'stream';
 import { createLogger, format, transports } from 'winston';
 
@@ -38,21 +40,18 @@ const errLogger = createLogger({
   ],
 });
 
-
-
-function requestLogger(req: Request, res: Response, next: CallableFunction): void {
-  if(req.body.password) req.body.password = 'Hidden';
+function requestLogger(req: Request, res: Response, next: NextFunction): void {
   const {method} = req;
   const {url} = req;
   const query = Object.getOwnPropertyNames(req.query).length ? `\nQuery: ${JSON.stringify(req.query)}` : '';
-  const params = Object.getOwnPropertyNames(req.params).length ? `\nParams: ${JSON.stringify(req.params)}` : '';
   const body = Object.getOwnPropertyNames(req.body).length ? `\nBody: ${JSON.stringify(req.body)}` : '';
-
-  logger.info(`Method: ${method} | Path: ${url} ${query} ${params} ${body}`);
+  logger.info(`Method: ${method} | Path: ${url} ${query} ${body}`);
   next();
 
-  finished(res, () => {
-    logger.info(`Response status code: ${res.statusCode}`);
+  finished(req, res, () => {
+    if(res.statusCode) {
+      logger.info(`Response status code: ${res.statusCode }`);
+    }
   })
 }
 
@@ -61,4 +60,12 @@ function errorLogger(err: Error, req: Request, res: Response, next: CallableFunc
   next();
 }
 
-export { requestLogger, errorLogger };
+// Used for logging uncaughtExceptions so the process.exit(1) doesn't stop writing to log files.
+function syncLogger(err: Error, title: string): void {
+const date = new Date();
+const dateNow = new Date(date.getTime() - (date.getTimezoneOffset() * 60000)).toISOString().slice(0, 19).replace('T',' ');
+writeFileSync(resolve(__dirname, '../logging/error-logs.log'), `${dateNow} ${title}: ${err.message}\n`, { flag: 'a'});
+process.stderr.write(`Error message: ${err.message}\n`);
+}
+
+export { requestLogger, errorLogger, syncLogger };
